+++ examples/ui-prompting-examples/src/utils/types.ts
@@ -6,7 +6,7 @@ import {
 } from '@sap-ux/fe-fpm-writer';
 import { PromptsType } from '@sap-ux/fe-fpm-writer/dist/prompts/types';
 import type { AddonActions } from '../addons/types';
-import type { DynamicChoices } from '@sap-ux/ui-prompting';
+import type { DynamicChoices, TranslationProperties } from '@sap-ux/ui-prompting';
 import type { Answers, CodeSnippet, SupportedGeneratorAnswers } from '@sap-ux/fe-fpm-writer';
 import type { I18nBundle } from '@sap-ux/i18n';
 
@@ -117,4 +117,5 @@ export interface CreateI18n {
     type: typeof CREATE_I18N_ENTRY;
     key: string;
     value: string;
+    properties?: TranslationProperties;
 }