+++ packages/cf-deploy-config-writer/src/cf-writer/index.ts
@@ -1,0 +1,2 @@
+export { generateAppConfig } from './app-config';
+export { generateBaseConfig } from './base-config';