+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @sap-ux/environment-check
 
+## 0.17.50
+
+### Patch Changes
+
+-   Updated dependencies [6275288]
+    -   @sap-ux/ui5-config@0.25.1
+    -   @sap-ux/project-access@1.28.5
+    -   @sap-ux/axios-extension@1.16.7
+
+## 0.17.49
+
+### Patch Changes
+
+-   Updated dependencies [5a68903]
+    -   @sap-ux/project-access@1.28.4
+    -   @sap-ux/axios-extension@1.16.7
+
 ## 0.17.48
 
 ### Patch Changes